issue 1: 空のセルへの対策など

+++ src/main/java/info/tmpz84/app/kassis/fileprocessor/doma/dao/UserDao.java
@@ -15,7 +15,7 @@ public interface UserDao {
     @Select
     UserEntity selectById(Integer id);
     @Select
-    int selectCount();
+    Long selectCount();
     @Insert
     int insert(UserEntity user);
     @Script
